Add dhcp_lease
