+++ dockers/docker-dhcp-server/lease_update.sh
@@ -3,7 +3,7 @@
 # it is to find running process dhcpservd.py, and send SIGUSR1 signal to this
 # process to inform it to update lease table in state_db (defined in dhcpservd.py)
 
-pid=`ps aux | grep 'dhcpservd.py' | grep -nv 'grep' | awk '{print $2}'`
+pid=`ps aux | grep 'dhcpservd' | grep -nv 'grep' | awk '{print $2}'`
 if [ -z "$pid" ]; then
     logger -p daemon.error Cannot find running dhcpservd.py.
 else
