+++ dockers/docker-dhcp-server/supervisord.conf
@@ -43,7 +43,7 @@ dependent_startup_wait_for=rsyslogd:running
 programs=dhcpservd,kea-dhcp4
 
 [program:dhcpservd]
-command=/usr/bin/python3 /usr/local/bin/dhcpservd.py
+command=/usr/local/bin/dhcpservd
 priority=3
 autostart=false
 autorestart=false
