+++ src/sonic-dhcpservd/.gitignore
@@ -1,0 +1,4 @@
+.eggs/
+build/
+dist/
+sonic_dhcpservd.egg-info/
