+++ src/sonic-dhcpservd/tests/test_dhcpservd.py
@@ -1,0 +1,2 @@
+def test_dhcpservd():
+    assert 1 == 1
